Move more architecture-specific code into architecture-specific files.
This makes it easier to see the high-level structure of the code in the
mod.rs files, and easier to find the target-specific code.
